Pin the version of the front end manifest used in production (see notes)
I was attempting to fetch the manifest dynamically during asset precompile
but as heroku runs asset precompile separately, any dynamic config I made
as part of that step was lost by the time the application started.

We could possibly generate this config file (config/spa_assets.yml) during
asset precompile (addig back the hook to precompile that was previously
in spa.rake) but is that actually a good idea? Is it super suprising or standard?

At the moment, the SPA assets have well known names at least for the javascript
entrypoint. However, I will probably reconfigure vite to its default behaviour so we
have a better story for managing versions and upgrades etc.