Adding subdomain constraints to all api + web routes (see notes)
SUBDOMAIN CONSTRAINTS

Subdomain constraints are a handy way of serving different stuff based on
matching 'subdomain' of the url. That means you can have the same routing logic
locally and in production - presumably with fiddling with /etc/hosts locally to
get local names working.

The only gotcha is the slightly annoying way that rails parses 'subdomains' and
'domains'. The behaviour depends on a tld_length setting (which we have set to
1). Presumably this dictates how many period-delimited tokens are determined to
be the 'domain' and how many are 'subdomain'. We have an issue that our two
supported domains: 'lml.live' and 'livemusiclocator.com.au' both have different
tld lengths. Which Rails does not support right now.

If we pick tld_length = 1:

- GOOD: api.lml.live is parsed as domain `lml.live` and subdomain `api`
- GOOD: potato.api.lml.live is parsed as domain `lml.live` and subdomain `potato.api`
- OK: www.livemusiclocator.com.au is parsed as domain `com.au` and subdomain `www.livemusiclocator`
- OK: beta.livemusiclocator.com.au is parsed as domain `com.au` and subdomain `beta.livemusiclocator`

If we pick tld_length = 2:

- BAD: api.lml.live is parsed as domain `api.lml.live` and no subdomain
- OK: potato.api.lml.live is parsed as domain `lml.live` and subdomain `potato`
- GOOD: www.livemusiclocator.com.au is parsed as domain `livemusiclocator.com.au` and subdomain `www`
- GOOD: beta.livemusiclocator.com.au is parsed as domain `livemusiclocator.com.au` and subdomain `beta`

This commit is based on option 1 - tld_length=1. The regex matching on subdomains just has to be a bit looser in certain cases.

This is all confusing and seems a bit fragile so I hope that introducing the routing specs for this work will help.

Other solutions to consider:

- change constraints to match the full host name
- add in a gem such as https://www.rubydoc.info/gems/tld_length to set the tld length dynamically
- redo our routes to all exist on the same subdomain

A few things to deal with or check heroku is ok with:

- Session management and cookies for admin portal
- Certs?
- Impact on caching with two subdomains
- Impact on cost serving things twice